Proposal: install everything with ocamlfind
Previously the META file was installed with `ocamlfind install` but
the binary artifacts were installed directly in the `STDLIBDIR`. If
using a system installation of OCaml (e.g. in `/usr/local/`) an
`opam install num` would install the `META` file in the opam root,
but the artifacts outside (e.g. in `/usr/local/lib/ocaml`). On some
systems the `STDLIBDIR` won't be writable and this will fail. On macOS
with homebrew it is writable but it fails intermittently in the following way:

- `opam init`
- `opam install num`
  - `num` depends on `ocamlfind`. The `ocamlfind` `configure` script
    https://github.com/ocaml/opam-repository/blob/master/packages/ocamlfind/ocamlfind.1.7.3/files/check-num-in-sitelib.patch
    looks for a `num.cmi` in the `STDLIBDIR`. The intention is to
    supply a `META` file for libraries which are distributed with
    the compiler. Since it doesn't find a `num.cmi` it doesn't generate
    a `META` file. The `opam install num` succeeds as expected.

- `rm -rf .opam`
  - at this point `num.cmi` exists in `STDLIBDIR`, left over from the
    previous installation

- `opam init`
- `opam install num`
  - The `ocamlfind` `configure` script detects a `num.cmi` in the
    `STDLIBDIR` and assumes that the compiler has `num` included.
    It therefore installs a `META` file for it. Unfortunately the
    `opam install num` runs `make install` which runs `ocamlfind install num`
    which fails since `num` is already installed.
    The `make uninstall` cleans up the `STDLIBDIR` so the next attempt
    would succeed.

This patch installs all the artifacts with `ocamlfind` into the opam
root. This is more like a normal OCaml library but is observably different
to before, where care was taken to install the library into the `STDLIBDIR`
to hide the difference between OCaml < 4.06 with `num` built-in and
OCaml 4.06+ with `num` as an external package.

This would fix [ocaml/opam-repository#10857]

Signed-off-by: David Scott <dave@recoil.org>