Fix incompatibility with OCaml 5.4
The type of long idents changed in OCaml 5.4.  Instead of building
longidents by hand, use the same trick as Zarith does: evaluate
`#install_printer` directives built as strings.

Note: the `Longident.unflatten` API would be great, but we cannot use it
since the `Longident` module is not visible in the REPL.

Fixes: #45